requirements.txt
numpy==1.21.5
pandas==1.4.2
Pillow==9.2.0
Plost==0.2.5
plotly==5.6.0
streamlit==1.12.0
